rename search tests to search intention tests
