+++ Makefile
@@ -6,7 +6,7 @@ test:
 	poetry run pytest -vvv -m "not search_test"
 
 test_not_vespa:
-	poetry run pytest -vvv -m "not vespa and not search_test"
+	poetry run pytest -vvv -m "not vespa and not search_intention"
 
-search_tests:
-	poetry run pytest -vvv -m "search_test"
+test_search_intentions:
+	poetry run pytest -vvv -m "search_intention"
